changed time() prototype from long time() to int time()
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@1018 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/platform/topcells/caba-vgmn-hello_world-iss2_lm32/soft/main.c
@@ -31,7 +31,7 @@
 #include "../segmentation.h"
 
 /* time function code is in stdio.c */
-extern long     time();
+extern int time();
 
 int fibo(int n) {
     if (n==0)
@@ -44,7 +44,7 @@ int fibo(int n) {
 int main(void) {
 
     printf("Hello from processor %d\n", procnum());
-    printf("Start Time %d\n",time ( (long *) 0));
+    printf("Start Time %d\n",time ( (int *) 0));
 
     printf("Fibo(1) = %d\n", fibo(1));
     printf("Fibo(2) = %d\n", fibo(2));
@@ -57,7 +57,7 @@ int main(void) {
     printf("Fibo(9) = %d\n", fibo(9));
     printf("Fibo(10) = %d\n", fibo(10));
 
-    printf("End Time %d\n",time ( (long *) 0));
+    printf("End Time %d\n",time ( (int *) 0));
 //    while(1);
     return 0;
 }
